Add tl;dr summary and example to README
Trying to understand this project with only outsider knowledge, I started reading the Source writeup (https://source.opennews.org/articles/how-npr-transcribes-and-fact-checks-debates-live/) and found the embedded tweet from @eads to be a really helpful high-level overview of the architecture that felt missing when I read through the README yesterday.  Add it inline to act as a road map for others reading the README for the first time.

Also add an examples section as I find it helps to see the finished product while also looking under the hood of an application.